pyproject.toml
[tool.poetry]
name = "Petproject"
version = "1"
description = "Petproject"
authors = ["MelekhSV"]

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^22.1.0"
make = "^0.1.6"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
pre-commit = "^2.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.black]
line-length = 100
target-version = ['py37', 'py38']
force-exclude = '''
(
    \.eggs
  | \.git
  | build
  | dist
  | venv
  | .venv
  | migrations
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip_gitignore = true
skip_glob = ['**/migrations/**', '**/.venv/**']
